Almacenamos siempre bien el orden de las cajas.
Antes daba problemas tanto si fallaban las validaciones como si las cajas ya existían (Rails no borra los registros existentes en una relación HABTM y mete los nuevos, sino que comprueba si ya existen).

+++ app/models/pagina.rb
@@ -6,11 +6,18 @@ class Pagina < ActiveRecord::Base
 
   has_many :sidebars
   has_many :cajas, :through => :sidebars
+  before_save :build_sidebar
 
-  def build_sidebar(caja_ids)
-    sidebars.destroy_all
-    caja_ids.each_with_index do |caja_id, index|
-      sidebars.build(:caja_id => caja_id, :orden => index + 1) unless caja_id.to_i.zero?
+  def build_sidebar
+    if @caja_ids
+      sidebars.destroy_all
+      @caja_ids.each_with_index do |caja_id, index|
+        sidebars.build(:caja_id => caja_id, :orden => index + 1) unless caja_id.to_i.zero?
+      end
     end
+  end
+
+  def ids_cajas=(caja_ids)
+    @caja_ids = caja_ids
   end
 end
